MMM-235 Add option to delete token frontend

+++ packages/backend/src/layers/functionality/tokens/delete-token.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { AccessDeniedError } from '~/common/errors/access-denied-error';
+import { ImportAttemptStatusType } from '~/storage/interfaces/create-monobank-token-import-attempt-dto.interface';
 import { ImportAttemptStorage } from '~/storage/services/import-attempt.storage';
 import { SpaceStorage } from '~/storage/services/space.storage';
 import { TokenStorage } from '~/storage/services/token.storage';
@@ -15,17 +16,33 @@ export class DeleteTokenService {
   constructor(
     private userStorage: UserStorage,
     private tokenStorage: TokenStorage,
+    private importAttemptStorage: ImportAttemptStorage,
   ) {}
 
   async delete({ tokenId, email }: IDeleteTokenArgs) {
-    const currentUserSpace = await this.userStorage.getSpaceByEmail(email);
-    const token = await this.tokenStorage.getTokenByTokenId(tokenId);
-    const isOwnSpace = currentUserSpace.id === token.space.id;
+    try {
+      const currentUserSpace = await this.userStorage.getSpaceByEmail(email);
+      const token = await this.tokenStorage.getTokenByTokenId(tokenId);
+      const isOwnSpace = currentUserSpace.id === token.space.id;
 
-    if (!isOwnSpace) {
-      throw new AccessDeniedError();
+      if (!isOwnSpace) {
+        throw new AccessDeniedError();
+      }
+
+      const tryFindImportAttempt = await this.importAttemptStorage.getByTokenId(
+        token.token,
+      );
+
+      const checkInProgressImport = tryFindImportAttempt.find(
+        (item) => item.status === ImportAttemptStatusType.InProgress,
+      );
+      if (checkInProgressImport) {
+        throw new Error('Statement import in progress');
+      }
+
+      return await this.tokenStorage.deleteToken(tokenId);
+    } catch (error) {
+      throw new InternalServerErrorException(error.message);
     }
-
-    return await this.tokenStorage.deleteToken(tokenId);
   }
 }
